daschlab/photometry.py: initial creation of this
